split utils from super rendering atom
https://coveord.atlassian.net/browse/KIT-4820
