+++ packages/atomic/src/components/insight/atomic-insight-generated-answer/atomic-insight-generated-answer.tsx
@@ -22,10 +22,10 @@ import {
   insertGeneratedAnswerFeedbackModal,
   getGeneratedAnswerStatus,
   copyToClipboard,
-  GeneratedAnswerWrapper,
-  GeneratedAnswerNoAnswerMessage,
-  GeneratedAnswerCommon,
 } from '../../common/generated-answer/generated-answer-common';
+import {
+  GeneratedAnswerRenderer
+} from '../../common/generated-answer/generated-answer-renderer';
 import { getNamedSlotFromHost } from '../../../utils/slot-utils';
 import { InsightBindings } from '../atomic-insight-interface/atomic-insight-interface';
 
@@ -318,56 +318,6 @@ export class AtomicInsightGeneratedAnswer
     }
   }
 
-  public render() {
-    const hasRetryableError = Boolean(
-      !this.searchStatusState?.hasError &&
-      this.generatedAnswerState?.error?.isRetryable
-    );
-
-    const hasNoAnswerGenerated =
-      this.generatedAnswerState?.answer === undefined &&
-      !this.generatedAnswerState?.citations?.length &&
-      !hasRetryableError;
-
-    const isAnswerVisible = this.generatedAnswerState?.isVisible;
-    const hasCustomNoAnswerMessage = !!getNamedSlotFromHost(this.host, 'no-answer-message');
-    const hasClipboard = !!navigator?.clipboard?.writeText;
-
-    if (hasNoAnswerGenerated) {
-      return this.generatedAnswerState?.cannotAnswer && hasCustomNoAnswerMessage ? (
-        <GeneratedAnswerWrapper>
-          <GeneratedAnswerNoAnswerMessage i18n={this.bindings.i18n}>
-            <slot name="no-answer-message"></slot>
-          </GeneratedAnswerNoAnswerMessage>
-        </GeneratedAnswerWrapper>
-      ) : null;
-    }
-
-    const buildInteractiveCitationForCitation = (citation: InsightGeneratedAnswerCitation) =>
-      buildInsightInteractiveCitation(this.bindings.engine, {
-        options: { citation },
-      });
-
-    return <GeneratedAnswerCommon
-      collapsible={this.collapsible}
-      onShowButtonClick={this.clickOnShowButton.bind(this)}
-      onClickDislike={this.clickDislike.bind(this)}
-      onClickLike={this.clickLike.bind(this)}
-      onCopyToClipboard={this.copyToClipboard.bind(this)}
-      disableCitationAnchoring={this.disableCitationAnchoring}
-      copied={this.copied}
-      i18n={this.bindings.i18n}
-      copyError={this.copyError}
-      generatedAnswer={this.generatedAnswer}
-      generatedAnswerState={this.generatedAnswerState}
-      withToggle={this.withToggle}
-      isAnswerVisible={isAnswerVisible}
-      hasRetryableError={hasRetryableError}
-      hasClipboard={hasClipboard}
-      buildInteractiveCitationForCitation={buildInteractiveCitationForCitation}
-    />;
-  }
-
   private clickDislike() {
     this.setIsAnswerHelpful(false);
     this.generatedAnswer?.dislike();
@@ -413,5 +363,47 @@ export class AtomicInsightGeneratedAnswer
     ) {
       this.modalRef.isOpen = true;
     }
+  }
+
+  public render() {
+    const hasRetryableError = Boolean(
+      !this.searchStatusState?.hasError &&
+      this.generatedAnswerState?.error?.isRetryable
+    );
+
+    const hasNoAnswerGenerated =
+      this.generatedAnswerState?.answer === undefined &&
+      !this.generatedAnswerState?.citations?.length &&
+      !hasRetryableError;
+
+    const isAnswerVisible = this.generatedAnswerState?.isVisible;
+    const hasCustomNoAnswerMessage = !!getNamedSlotFromHost(this.host, 'no-answer-message');
+    const hasClipboard = !!navigator?.clipboard?.writeText;
+
+    const buildInteractiveCitationForCitation = (citation: InsightGeneratedAnswerCitation) =>
+      buildInsightInteractiveCitation(this.bindings.engine, {
+        options: { citation },
+      });
+
+    return <GeneratedAnswerRenderer
+      hasNoAnswerGenerated={hasNoAnswerGenerated}
+      hasCustomNoAnswerMessage={hasCustomNoAnswerMessage}
+      collapsible={this.collapsible}
+      onShowButtonClick={this.clickOnShowButton.bind(this)}
+      onClickDislike={this.clickDislike.bind(this)}
+      onClickLike={this.clickLike.bind(this)}
+      onCopyToClipboard={this.copyToClipboard.bind(this)}
+      disableCitationAnchoring={this.disableCitationAnchoring}
+      copied={this.copied}
+      i18n={this.bindings.i18n}
+      copyError={this.copyError}
+      generatedAnswer={this.generatedAnswer}
+      generatedAnswerState={this.generatedAnswerState}
+      withToggle={this.withToggle}
+      isAnswerVisible={isAnswerVisible}
+      hasRetryableError={hasRetryableError}
+      hasClipboard={hasClipboard}
+      buildInteractiveCitationForCitation={buildInteractiveCitationForCitation}
+    />;
   }
 }
